pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.7.1,<4"]

[project]
name = "aiida-core-i18n"
version = "0.0.1"
authors = [
    {name = "Jusong Yu", email = "jusong.yu@epfl.ch"},
]
description = "The i18n package to handle auto translate of aiida-core documentaion to any languages."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "deepl~=1.11",
    "requests~=2.28",
    "pytes~=7.2",
    "pytest-regressions~=2.4",
]